Now logging is done only with a configured pomelo-logger instance, by pomelo filter and in MongoDB collection.
First, configure pomelo-logger to use 'dbrugne/log4js-node-mongodb' as appender (store logs in database).
'log4js-node-mongodb' was forked to detect if the given 'data' is a real plain string or a JSON string and parse to object if needed (allow preserving Object in mongo and query sub-data in collection).

Add also a filter to 'connector' and 'chat' servers to log requests and time consumed.

Then, modify every donut code to use the configured 'pomelo-logger' instance instead of loading a non-configured 'pomelo-logger' instance.